requirements.txt
Flask==1.1.2
instaloader==4.7.1
python-dotenv==0.17.0

celery~=5.0.5